app/protobufs/spark_connector/requirements.txt
grpc-interceptor >= 0.11.0
grpcio-tools >= 1.30
dnspython >= 2.1.0
google-cloud-dataproc
google-cloud
google-cloud-storage
# pyspark >= 0.0.0